// src/app/services/models/ProjectToReturnTotal.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Image } from '../models/image';
import { Pdf } from '../models/pdf';
import { Video } from '../models/video';
import {EvaluationToReturn} from "./evaluation-to-return";
export interface ProjectToReturnTotal {
  description?: string;
  domaine?: 'Web' | 'CyberSecurity' | 'Ai_Models';
  image: Array<Image>;
  isOpen?: boolean;
  isVisible?: boolean;
  meanEvaluations: number;
  name?: string;
  professorFullName?: string;
  report: Pdf;
  requestsTotal?: number;
  studentFullName?: string;
  totalEvaluations: number;
  video?: Video;
  evaluationToReturns:Array<EvaluationToReturn>;
  studentId?:number;
  professorId?:number;
  professorImage?:string,
  studentImage?:string
}